+++ OsuDiffCalc/FileProcessor/AnalyzerObjects/DifficultyRating.cs
@@ -120,13 +120,13 @@
 			return GetNormalizedSeries(points);
 		}
 
-		public void AddJump(int timeMs, float difficulty) => Add(timeMs, difficulty, _jumps);
-		public void AddStream(int timeMs, float difficulty) => Add(timeMs, difficulty, _streams);
-		public void AddBurst(int timeMs, float difficulty) => Add(timeMs, difficulty, _bursts);
-		public void AddDouble(int timeMs, float difficulty) => Add(timeMs, difficulty, _doubles);
-		public void AddSlider(int timeMs, float difficulty) => Add(timeMs, difficulty, _sliders);
+		public void AddJump(double timeMs, float difficulty) => Add(timeMs, difficulty, _jumps);
+		public void AddStream(double timeMs, float difficulty) => Add(timeMs, difficulty, _streams);
+		public void AddBurst(double timeMs, float difficulty) => Add(timeMs, difficulty, _bursts);
+		public void AddDouble(double timeMs, float difficulty) => Add(timeMs, difficulty, _doubles);
+		public void AddSlider(double timeMs, float difficulty) => Add(timeMs, difficulty, _sliders);
 
-		private void Add(int timeMs, float diff, SeriesPointCollection dest) {
+		private void Add(double timeMs, float diff, SeriesPointCollection dest) {
 			lock (_pointsLock) {
 				float xValue = (float)(timeMs / 1000.0);
 				_allSeriesXValues.Add(xValue);
