+++ OsuDiffCalc/FileProcessor/BeatmapObjects/BreakSection.cs
@@ -1,6 +1,6 @@
 ﻿namespace OsuDiffCalc.FileProcessor.BeatmapObjects {
 	/// <summary> A break in gameplay </summary>
 	/// <inheritdoc cref="BeatmapObject"/>
-	record BreakSection(int StartTime, int EndTime) : BeatmapObject(StartTime, EndTime) {
+	record BreakSection(double StartTime, double EndTime) : BeatmapObject(StartTime, EndTime) {
 	}
 }
